+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @stackframe/stack-dashboard
 
+## 2.5.24
+
+### Patch Changes
+
+- Various bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.24
+  - @stackframe/stack@2.5.24
+  - @stackframe/stack-emails@2.5.24
+  - @stackframe/stack-ui@2.5.24
+
 ## 2.5.23
 
 ### Patch Changes