+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/oauth-mock-server
 
+## 2.5.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.24
+
 ## 2.5.23
 
 ### Patch Changes