+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-emails
 
+## 2.5.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.24
+  - @stackframe/stack-ui@2.5.24
+
 ## 2.5.23
 
 ### Patch Changes